move function + fix issue
