+++ apps/backend-functions/src/internals/invitations/events.ts
@@ -1,12 +1,11 @@
 import { InvitationOrUndefined, InvitationDocument } from "@blockframes/invitation/+state/invitation.firestore";
-import { wasCreated, wasAccepted, wasDeclined } from "./utils";
+import { wasCreated, wasAccepted, wasDeclined, hasUserAnOrgOrIsAlreadyInvited } from "./utils";
 import { NotificationDocument, NotificationTypes, OrganizationDocument, PublicUser } from "../../data/types";
 import { createNotification, triggerNotifications } from "../../notification";
 import { getUser } from "../utils";
 import { createDocumentMeta, createPublicInvitationDocument, getAdminIds, getDocument } from "../../data/internals";
 import { EventDocument, EventMeta } from "@blockframes/event/+state/event.firestore";
 import * as admin from 'firebase-admin';
-import { hasUserAnOrgOrIsAlreadyInvited } from '../../invitation';
 
 /**
  * Handles notifications and emails when an invitation to an event is created.
@@ -33,19 +32,14 @@ async function onInvitationToAnEventCreate(invitation: InvitationDocument) {
   const recipients: string[] = [];
   if (!!invitation.toUser) {
     /**
-     * @dev We wants to send this email only if user have an orgId. If not, this means that he already received an
+     * @dev We wants to send this email only if user have an orgId and a validated account. If not, this means that he already received an
      * email inviting him along with his credentials.
     */
     const user = await getDocument<PublicUser>(`users/${invitation.toUser.uid}`);
-    const hasOrgOrOrgInvitation = await hasUserAnOrgOrIsAlreadyInvited([...invitation.toUser.email]);
-    console.log('!!!!!!!!!! hasOrgOrOrgInvitation', hasOrgOrOrgInvitation);
+    const hasOrgOrOrgInvitation = await hasUserAnOrgOrIsAlreadyInvited([invitation.toUser.email]);
     if (!user.orgId && !hasOrgOrOrgInvitation) {
       console.log('Invitation have already been sent along with user credentials');
       return;
-    }
-
-    if (!user.orgId && !!hasOrgOrOrgInvitation) {
-      recipients.push(invitation.toUser.uid);
     }
 
     recipients.push(invitation.toUser.uid);
